[IMP] base/saas~16.5: clean attrs modifiers
We no longer allow for `attrs` attribute in fields. We rely instead on
inline attributes with general Python expressions. We improve here the
previous script that handles this. The main goal of the rewrite is to
simplify the migration of the attrs usage. We don't deal with standard
views since they are supposed to be correct after the upgrade.

We need to use AST because of top level if expressions or boolean
expressions that are used as `attrs` dict values. `safe_eval` won't work
in such cases.

After converting all base archs (en_US lang) we need to make an extra
pass for other languages because the translations are not prefect. This
is especially recurrent for <span> elements for which the attributes are
not merged correctly from the base arch into the translations.

Some tests were adapted to match the slight differences in output.
* Do not optimize. For example cases like `xxx == True` won't be
  optimized into just `xxx`.
* Method c14n orders attributes, which makes better for comparing
  results.

Some extra tests for inactive/error views.

The general process to update the views is as follows:
1. Deactivate all views.
2. Iterate views top-down (i.e. views that are applied later in their
   hierarchy should be processed later).
3. Update views taking into account their combined arch.
4. If a view was originally inactive or it has errors in its arch, then
   fix it without combined arch.
5. Reactivate the view. This is important to guarantee that next views
   will see the changes applied by this one.

Part-of: odoo/upgrade#5090
Co-authored-by: Gorash <chm@odoo.com>