[IMP] account_reports/16.0: Support sum_if_(pos|neg)_groupby
The `sum_if_(pos|neg)_groupby` aggregation expression was removed in
Odoo 16, replaced by C/D (Credit/Debit) flags at the end of the account
code.

This commit is to make sure to migrate such expressions as aggregators
and not aggregation terms. Note, that migrated reports making use of such
expressions will not open (since it is no longer supported). However,
this will make it easier for either customers or specific scripts to fix
the report back.

closes odoo/upgrade#5056

Signed-off-by: Christophe Simonis (chs) <chs@odoo.com>