[IMP] hr_appraisal_survey/saas~16.3: update record rule
When users with `User` access rights tries to access a survey form,
they gets an Odoo server error.

The error is caused by a record rule that calls on a field that doesn't
exist in the survey model.

The purpose of this PR is to update that record rule, after
it has been fixed.

see odoo/enterprise#46093
task-3474933

closes odoo/upgrade#5061

Signed-off-by: Christophe Simonis (chs) <chs@odoo.com>