[FIX] account_reports/16.0: Update `sum_if_(pos|neg)_groupby` domains
[OPW-3462046](
https://accounts.odoo.com/web#id=3462046&menu_id=4720&cids=1&action=5050&model=project.task&view_type=form)
[UPG-1090459](
https://upgrade.odoo.com/web#cids=1&id=1090459&model=upgrade.request&view_type=form&menu_id=)

`sum_if_(pos|neg)_groupby` is no longer supported in Odoo 16.
However, domains associated with such aggregators are not properly
translated. The result is that reports with at least one line using one
such aggregator will fail to open after the upgrade.

This adds a proper migration to the Odoo 16 way, for domains associated
to the aforementioned groupby aggregators.

closes odoo/upgrade#5065

Signed-off-by: Christophe Simonis (chs) <chs@odoo.com>