[IMP] hr_*/saas~16.5: generalize activity plan
We generalize the activity plan from a feature only applicable to hr.employee
to any model implementing activity mixin. The specificity of this upgrade is
that we have to move the data formerly defined in the hr module (hr.plan and
hr.plan.activity.type) to the new generic model defined in a parent module:
mail (mail.activity.plan and mail.activity.template). The model fields are
almost identical with some small changes (ex.: a plan is now linked to a model
while it was previously implicitly linked to hr.employee).

The upgrade is done in the hr module to migrate the former specific models to
the generic one including fields added in sub modules. A test has been added in
the hr upgrade and in all sub modules where the former/new models were/are
modified (hr_fleet and hr_contract_sign).

See odoo/odoo#137969
See odoo/enterprise#48599

Task-3390865

closes odoo/upgrade#5243

Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>