[IMP] hr_{work_entry_}holidays/saas~16.5: upgrade
Compute new stored fields resource-calendar_id and number_of_hours.
number_of_hours needs to be computed explicitly because the
automatic computation of it would trigger the recomputation of
number_of_days as well, which could override user modifications.

It would have been great to do this computation in SQL, but it would be
very complex to do so.

See odoo/odoo#119317
See odoo/enterprise#40676

closes odoo/upgrade#5039

Signed-off-by: Christophe Simonis (chs) <chs@odoo.com>