[IMP] sign/saas~16.5: remove refusal option and signature order group
In sign it is possible to allow specific sign requests to be refused. However,
this is currently an option and not all sign requests are refusable by default.
As in 16.5, we allow this by default, there's no need to keep the
refusal_allowed fields in the wizard and in the signature.request model.

Sign also allows setting an order of signature when sending a sign request.
However, this feature is currently opt-in in the settings of sign. The related
PR changes this to have this feature available by default, so we can also get
rid of the group related to enabling this feature.

Related PR: https://github.com/odoo/enterprise/pull/46339

task-3457176

closes odoo/upgrade#5079

Related: odoo/enterprise#46339
Signed-off-by: Christophe Simonis (chs) <chs@odoo.com>