[IMP] {mass_mailing_,}sms/saas~16.5: adapt for SMS v3
UUIDs are added to sms records as identifiers to
communicate with IAP via API and webhooks for delivery reports
(to update mail_notifications and mail_trace).

A new table is added to link notifications and traces
to sms so sms can be deleted early (when delivery reports arrive).
This means removing the foreign key constraint, but the sms id
should be kept they are used in short links sent to users.

Bonus: Removing the sms.api model to not be part of the environment.
Task-3488393

Task-2560666
See odoo/odoo#133392
See odoo/enterprise#46996

closes odoo/upgrade#5164

Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>