[FIX] helpdesk/saas~15.1:correctly handle the priorities in 'helpdesk'
The strings of priority 0 and 1 has changed here
(https://github.com/odoo/enterprise/pull/20228/files)
Because of this, all the tickets with mediom priority ("1") changes to low
priority ("0"). The values of string has for priority 0 has been changed to
all to low priority and for priority,it has been changed from low priority to
medium priority. The same changes were adapted in v15 as well.

closes odoo/upgrade#5100

Signed-off-by: Christophe Simonis (chs) <chs@odoo.com>