[IMP] account_followup/saas~16.5: recreate reconcile index
To improve the UX when reconciling, move lines matching the value of the bank
transaction will appear first.
To achieve this, a custom ORDER BY is used. This could have a significant
performance impact on dbs with a large number of unreconciled transactions.
In order to minimise the performance hit, we try to use an existing index.
This change recreates the index to accomodate the query executed by the embedded
list view.
The index was created by account_followup and is now moved to account

See odoo/odoo#135155
See odoo/enterprise#45816

closes odoo/upgrade#5146

Signed-off-by: Alvaro Fuentes Suarez (afu) <afu@odoo.com>