[FIX] util/records: fix update from xml of menus
Updating menus that are a child of another menu and that
do not have the field `parent_id`, would result in the loss
of the link between the menu and its parent.

In order to retain this link, we obtain the value for `parent_id`
and then explicitly inject this field to the source xml.

closes odoo/upgrade#5163

Signed-off-by: Christophe Simonis (chs) <chs@odoo.com>