[IMP] l10n_lu_reports/saas~16.5.1.0: refactor annual tax report
Currently, the LU annual VAT declaration is a large custom model, because
it was not possible to add manual values in the regular reports,
default values, and sections.
Now it is finally possible to get rid of that custom report and
use the account report engine to create the Annual VAT declaration.
- Migration has to be done in post_migrate,
because we need the new report to be created
- That means that l10n_lu_reports_annual_vat_2023 module
can only be deleted after, if possible
- Report engine does not support separate codes for each column,
a code can only be added on a line. We need a fields_mapping
for the multi-column lines to target the right expression.
- Year and company_id fields are moved from l10n_lu.yearly.tax.report.manual
model to l10n_lu_reports.report.appendix.expenditures
- l10n_lu_yearly_tax_report_manual has many columns,
so it's easier to fetch columns names and data types
- l10n_lu.yearly.tax.report.manual and related views and records
are deleted

See:
odoo/odoo#135783
odoo/enterprise#47525

closes odoo/upgrade#5198

Signed-off-by: Alvaro Fuentes Suarez (afu) <afu@odoo.com>