[FIX] base/0.0.0 : Check if failing field was removed in current view
When a studio view removes a field it may cause errors in some other parts
of the view hierarchy. There was a previous case for when a field was not found
that relied on transforming the change from <field position=replace> into
<xpath position=replace>. The issue with this approach is that it may happen
that the failing element is another trying to anchor onto the same field.
Consider the following example.

A primary view A uses a field as anchor.

  <field name='x' position='after'>...

A primary view B with parent A removes the field

  <xpath expr="//field[@name='x']" position='replace'/>

Note, maybe A got the new part that needs the anchoring field
during the upgrade.

Now when rendering A there is no failure. But when we request the rendering of B
we get the error <field name='x' ... not found. We cannot find any element of
the form <field name='x'... in B because the issue is in A. Thus the case that
transforms this into an xpath won't be triggered. We must deal with this extra
case where the failure is already an xpath element. The solution is similar as
already done: instead of removing we hide the field.

closes odoo/upgrade#5179

Signed-off-by: Alvaro Fuentes Suarez (afu) <afu@odoo.com>