[FIX] sale/0.0.0: remove uniq constraint
The unique constraint on name,company_id sale_order_name_uniq
was removed in version 9.0 in 2015, but not removed in
upgrade scripts until 2022.
In many failed requests, either some subscriptions have the same
name, or they have the same name as a sale order. When moving
the subs to sale orders, the unique constraint fails the upgrade.

closes odoo/upgrade#5182

Signed-off-by: Christophe Simonis (chs) <chs@odoo.com>