[IMP] l10n_es/saas~16.5: move simplified partner and preset new field
The simplified partner was moved from l10n_es_edi_sii to l10n_es.
We also add a field that is True when we use the simplified partner
by default and so we set the old account moves to true that used
that simplified partner.

Linked to https://github.com/odoo/odoo/pull/136164

closes odoo/upgrade#5191

Related: odoo/odoo#136164
Signed-off-by: Christophe Simonis (chs) <chs@odoo.com>