[ADD] account{,_avatax}/saas~17.1: deprecate Taxcloud
The accounting settings has a checkbox to install Taxcloud. Taxcloud
will be deprecated in Odoo 17 so this checkbox is removed. Because of
this we no longer need this field.

On top of that Avatax is auto-installed for the appropriate countries
so we removed that field from the accounting settings as well.

Community PR: https://github.com/odoo/odoo/pull/136894
Enterprise PR: https://github.com/odoo/enterprise/pull/48045

task-3516826

closes odoo/upgrade#5199

Related: odoo/odoo#136894
Related: odoo/enterprise#48045
Signed-off-by: Christophe Simonis (chs) <chs@odoo.com>