[IMP] account_*/saas~16.{3,5}: new deferred fields
*accountant,asset

In the related PR below we split the existing field
`generate_deferred_entries_method` on `res.company` in two new fields.

This upgrade script makes sure to copy the value of the old field in the
new fields and remove the old one. It also adapt the previous script in
`account_asset` to calculate the values of the two new fields instead
of the old field when upgrading from an older version.

See https://github.com/odoo/enterprise/pull/48022

task-3516660

closes odoo/upgrade#5200

Related: odoo/enterprise#48022
Signed-off-by: Christophe Simonis (chs) <chs@odoo.com>