[FIX] hr_payroll/0.0.0: fail on unlink with a message
Fail unlink of hr.salary.rule with an explicit message

When `appears_on_payroll_report` parameter is passed to write method, it creates
a field on hr_payroll_report when this flag is true and removes when its false.

This function:
https://github.com/odoo/enterprise/blob/16.0/hr_payroll/models/hr_salary_rule.py#L172

in hr_payroll_reports tries to remove fields but it can't/doesn't
call `util.remove_field` method so it fails.

closes odoo/upgrade#4984

Signed-off-by: Alvaro Fuentes Suarez (afu) <afu@odoo.com>