[IMP] appointment/saas~16.5: rename category and move anytime
Rename the 'website' category of appointment type to
'recurring'. This is done as with the arrival of
the new 'punctual' appointment type category,
the 'website' category isn't the only one available
on the website anymore.

Rename the xmlid of the action opening the Invitations
menu to remove the 'anytime' part as anytime appointment
types are now displayed in the Appointments menu.

See odoo/enterprise#43298

Task-3109254

closes odoo/upgrade#5170

Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>