[IMP] hr_homeworking/saas~16.5: remove fields and reset exceptions
https://github.com/odoo/odoo/pull/137502 changes how we handle homeworking
records. Some fields are deleted because we simplified how we create
exceptions in homeworking, some wizards are no longer required. This commit
adapts the hr.employee.location model, resets the employee locations previously
created because we don't need them anymore. Most cases should be handled by
having a default worklocation set on a day of the week. Also, due to the big
changes in the exception creation, keeping the old ones doesn't make a lot of
sense because they were created with other fields in mind.

task-3527196

closes odoo/upgrade#5260

Related: odoo/odoo#137502
Signed-off-by: Alvaro Fuentes Suarez (afu) <afu@odoo.com>