[FIX] base/0.0.0 : Modify the re pattern for group
This commit modify the search pattern for the error
which are caused when field1 is trying to access data
of field2 as it is used in domain of field1 then the error
pattern contain whole node instead of just the field name.
So pattern has to modified to extract name of the field
from the node.

closes odoo/upgrade#5247

Signed-off-by: Alvaro Fuentes Suarez (afu) <afu@odoo.com>