[IMP] hr_holidays,mail/saas~16.5: improve activity view
We add the active field in the mail.activity model (default true). The goal is
to support keeping "done" activity. With this change, when an activity is
marked as "done" and this feature is enabled on the related activity type
(disabled by default), the activity will be archived instead of being deleted.
We have also modified the code to display archived activity in the activity
view.

Delay count of "Time Off Approval" activity type (mail_act_leave_approval) is
now used to compute the activity deadline when requesting a leave. We set it to
15 (instead of 0) meaning that the deadline of the request activity will be set
15 days before the requested leave (if not before today). This change will help
the approver because those activities are now more or less sorted by leave
start date.

See odoo/odoo#138135
See odoo/enterprise#48685

Task-3390865

closes odoo/upgrade#5264

Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>