[FIX] website/16.0.1.0: fix translate and edit_master button
on changing language the 'translate' and 'or edit_master' button is not
visible and in js side  using html_data to show these button  if their
is no 'translatable' and 'editable'  set as true it will not show
these button and when we upgrade database from older version to 16.0
version standard website `Main layout` has been updated but related
Cowed view has not been properly updated. So need to add this element
manually here on `Main layout`  view.

UPG-1146390
OPW-3526170

closes odoo/upgrade#5296

Signed-off-by: Alvaro Fuentes Suarez (afu) <afu@odoo.com>