[FIX] loyalty/saas~15.3: Fix Missing discounted specific_products
The list of specific_products(stored in m2m table
'coupon_reward_product_product_rel') on which the discount
is applied specifically in promotion as well as coupon_programs,
does not get migrated moving to v16. As the the data
of dropped m2m table is not inserted through upgrade script. We
insert the data of dropped table in destined table
('loyalty_reward_product_product_rel') to ensure data integrity.
Moreover to it, the script for loyalty_program
(having discount on specific_products) exists. Similarly for covering
missing data from promotion and coupon_programs, same approach
is followed in this script.

opw-3557591
upg-1153049

closes odoo/upgrade#5311

Signed-off-by: Alvaro Fuentes Suarez (afu) <afu@odoo.com>