[IMP] hr/saas~16.5: reassign activities
When scheduling plans on employees, activities may be assigned to people who
don't have access to the record. It was managed locally in HR through a dummy
model aggregating activities (from version 16.3). We have removed that model
and allowed user to see the activities they are assigned to. To keep those
activities, we reassign them to hr.employee instead of the dummy model.

See odoo/odoo#139917

Task-3562098

closes odoo/upgrade#5339

Signed-off-by: Alvaro Fuentes Suarez (afu) <afu@odoo.com>