[FIX] base/0.0.0: handle group error resolution in child views
If a group is applied in child view using xpath/field element like :
<xpath/field expr="anything"/name="field_name" position="attributes">
    <attribute name="groups">group_name</attribute>
</xpath/field>

So in this kind of scenario we will have error like :

Field "group_guarded_field_name" used in
domain/attrs/context xyz is restricted to the group(s) group_name.'

The element causing this error does not exist in this view; it is located in
the parent view. To resolve this, we add an invisible field
using 'group_guarded_field_name' and 'group_name'.

Example parent view :

<data>
<field name="x"/>
<field name="y" attrs={'invisible':[('x', '=', 0)]}/>
</data>

child view:

<xpath expr="//field[@name = 'x']" position="attributes">
    <attribute name="groups">xyz</attribute>
</xpath>

Now when we load child view, this view will try to apply group 'xyz' to 'x'
field but by doing this field y will not be able to access field x
and raise error :

`Field \'x\' used in attrs ({'invisible':[('x', '=', 0)]}) is restricted to the
group(s) xyz.`

To tackle this situation we take group name, field_name from error message and
add an invisible field accordingly so after the fix it will be like

child view:

<xpath expr="//field[@name = 'x']" position="before">
   <field name='x' invisible='1'/>
</xpath>

<xpath expr="//field[@name = 'x']" position="attributes">
    <attribute name="groups">xyz</attribute>
</xpath>

here the expr is taken of the node which is adding this group and not created
from error message because it is possible that the group is applied
to whole page or div.

Part-of: odoo/upgrade#5337