[FIX] base/0.0.0: always restore standard views
We cannot rely on the noupdate flag. For example when a view is part of
a noupdate=0 block at the start of the upgrade, but it's on a noupdate=1
on the target version then the noupdate flag would never be set. Still
the view must be updated from the xml if it is wrong.

For a more concrete example, the view on
https://github.com/odoo/odoo/blob/8830af8/addons/website_forum/views/website_pages_views.xml#L4
starts in noupdate=0, but after rename it ends up in a noupdate=1 block
https://github.com/odoo/odoo/blob/21e551f/addons/website_forum/views/forum_post_views.xml#L4
This views needs to be updated because of the change of `attrs` in 17.0

Part-of: odoo/upgrade#5090