[FIX] l10n_ke_hr_payroll/16.0: move removal to post script
Removing of the record in pre script was causing fk violation
because the record l10n_ke_employees_salary_paye, which was
referencing this record was not updated by ORM.

The category is removed from the data files but since the ORM
has not yet updated the record in pre, it is still in-use.
By removing the unused record in post we have higher chances
that the record would actually be removed.

closes odoo/upgrade#5341

Signed-off-by: Alvaro Fuentes Suarez (afu) <afu@odoo.com>