[FIX] hr_homeworking/saas~16.5: remove default worklocation
See https://github.com/odoo/odoo/pull/141209
Recently we introduced various changes to homeworking. One of those was the
default work location being set as "office". However, it was later decided to
keep the default location as false. This commit removes the part of the
upgrade script that sets the default location as office.

task-3584193

closes odoo/upgrade#5351

Signed-off-by: Christophe Simonis (chs) <chs@odoo.com>