[IMP] stock/saas~16.5: mix of qty_done and reserved_qty
Following pull request [1]
The purpose is to merge both qty_done and product_qty on stock.move.line
qty_done is always an input from the users while product_qty is mainly
fill by the system and represent the quantity reserved for the picking.

The purpose of the modificiation is explained in the community PR
For the upgrade we choose to keep the qty_done and fill the quantity
column with them. Also each move with qty_done is marked as picked to
prevent modification later by the system. In order to achieve that, we
drop the reservation on exisiting databases. They are added back later
due to the scheduler action

[1] #5310

see odoo/odoo#137864
see odoo/enterprise#48709

closes odoo/upgrade#5366

Signed-off-by: Christophe Simonis (chs) <chs@odoo.com>