[FIX] uom/0.0.0: match xmlid if not exists
The matching of ref units with their xmlids was
originally meant to create an xmlid for existing
units if the xmlid was removed.
The edit of the util method lead to change the record
that the xmlid refers to if it was different than the
original. which is not a desirable effect for uoms.

closes odoo/upgrade#5373

Signed-off-by: Alvaro Fuentes Suarez (afu) <afu@odoo.com>