[FIX] base/saas~16.5: reset all modifiers on attrs reset
When a spec sets `attrs` it must reset the extra values of the target spec.
For example if we have
```xml
<field name="x" attrs="{'readonly':1}"/>
```
and in another view we target that field
```xml
<field name="x" position="replace" attrs="{'invisible':1}"/>
```
Then we need to keep only `invisible` modifier, thus instead of
```xml
<field name="x" readonly="1" invisible="1"/>
```
our output must be
```xml
<field name="x" invisible="1"/>
```

This is especially relevant for the fake element strategy that we use
for migrating the invalid modifiers when for `<attribute/>` elements.

Another issue fixed here, that pointed out the above issue, is that when
we reset the `attrs` attribute in an `<attribute/>` element we must set
an empty dict as new value, not an empty string. Otherwise we cause a
parse error in the `ast.parse` call.

Part-of: odoo/upgrade#5388