[FIX] account_reports/16.0: better matching
The lookahead and lookbehind were matching only if the next/prev char
was a space or the end/start of the string. However in some cases we may
have no space between operators (e.g. `+`, `-`) and terms. Getting to
something like `A.balance+sum.balance`. Thus `\W` is a better match than
`\s`.

closes odoo/upgrade#5396

Signed-off-by: Alvaro Fuentes Suarez (afu) <afu@odoo.com>