[FIX] base/0.0.0: child remove field accessed by another field
when we remove a field using studio which is being used by another field can
cause issue for group restriction.
Parent :
<field name=x>
<field name=x group='abc'>
<field name=y attrs={'invisible':[('x', '=', 1)]}>

child :
<xpath expr"field[@name="x]" position='replace'>
    <field name=f1 >
</xpath>

Combined:
<field name=f1 >
<field name=x group=abc>
<field name=y attrs={'invisible':[('x', '=', 1)]}>

we get error like :
Field \'x\' used in attrs ({\'invisible\': [[\'x\',\'=\',1]]})
is restricted to the group(s) abc

This error is coming because the attrs used in y try to access the x field which
is guarded by group abc This issue is coming from the parent but it is coming
because of the child node trying to remove the field x from the parent.

To fix this problem, we can conceal the field instead of deleting it.

child after patch :
<xpath expr="//field[@name='x']" position="attributes">
    <attribute name="invisible">1</attribute>
</xpath>
<xpath expr="//field[@name='x']" position="after">
    <field name=f1 >
</xpath>

combine after patch:
<field name=x invisible=1>
<field name=f1 >
<field name=x group=abc>
<field name=y attrs={'invisible':[('x', '=', 1)]}>

closes odoo/upgrade#5337

Signed-off-by: Alvaro Fuentes Suarez (afu) <afu@odoo.com>