[FIX] base/saas~16.5: fix act server convertion
During the refactor of server actions and their object lines,
the lines with type=value are expected to have an intiger value
but sometimes we find strings, which is allowed in Odoo and is
easily reproduceable. When the action is run, the value is ignored
when it is of a different type.

closes odoo/upgrade#5404

Signed-off-by: Christophe Simonis (chs) <chs@odoo.com>