[IMP] planning/saas~17.3: remove duration planning
This commit removes the duration field in the planning slot template as
the logic is changed.
Before removing the fields, this commit also computes the new field
`duration_days` and `end_time` since now those fields are stored and be
editable by the user.
This computation will just use the duration to add to the start time set
on the template to find the end_time and the duration_days.

related-PR: odoo/enterprise#52712
task-3624669

closes odoo/upgrade#5438

Signed-off-by: Christophe Simonis (chs) <chs@odoo.com>