[IMP] documents{_account}/saas~17.1: show embedded PDF inside XML
Purpose
=======
In accounting, some XML invoice documents have PDF embedded in base 64.
We want to show a preview of the embedded PDF, and also to view that
PDF when we click on the preview.

To ease the generation of the thumbnail, we added a new selection
option "to_generate", so we need to parse the XML only once.

Only the PDFs are set to "client_generated" if needed, because there should
not be a lot of PDFs without thumbnails (they are generated when we upload
the PDF). About the XML, there can be a lot of XML with embedded PDFs.
and we will need to parse them all during the migration, and the first
user opening documents will generate all thumbnails at once, so we
decided to let the flag to false for those files.

Task-3537562
See odoo/enterprise/pull/49111

closes odoo/upgrade#5437

Related: odoo/enterprise#49111
Signed-off-by: Christophe Simonis (chs) <chs@odoo.com>