[IMP] sale_purchase_inter_company_rules/saas~17.2: Link lots
The feature changes from being able to copy a lot name from a company to
another to being able to properly link the now cross-company lots from
one delivery to a reception, allowing to keep the traceability among
companies.
Changes the name of the field to better reflect that.

See odoo/enterprise#52882

Part-of: odoo/upgrade#5478