[IMP] sale_subscription/saas~17.1: archived pricelists from the products
The product will be removed from the pricelist when archived,
but it won't be removed from the product when archived.
This commit has fixed the issue. Once the pricelist is achieved,
it will be removed from the product and once the pricelist
is unachieved, it will be added to the product.

Related: odoo/enterprise#50262

task-3573491

closes odoo/upgrade#5451

Signed-off-by: Alvaro Fuentes Suarez (afu) <afu@odoo.com>