[FIX] sale/saas~16.2: fix join in `amount_to_invoice` computation
Currently, the computation of `amount_to_invoice` for a partially
invoiced sale order doesn't work.

There is a mistake in the migration script, when joining the
`sale.order.line` and the `account.move.line` using their relation
table, the order line ids are joined with the move line ids and
vice-versa.

The fix is to join the order line with the order line ids from the
relation table.  Same for the move lines.

opw-3625908

closes odoo/upgrade#5452

Signed-off-by: Alvaro Fuentes Suarez (afu) <afu@odoo.com>