[FIX] account/saas~17.3: new field `invoice_currency_rate` on move
In the related community PR (see below) a new computed and stored field
`invoice_currency_rate` is added to account.move. In this commit the
respective column is added to the DB and its values are computed
(differently from the new compute method).

The new field tracks the exchange rate at invoice date.
This rate is also displayed on the form view.

One of the reasons for adding the field was that previously the
individual lines of an invoice could use different currency rates.
Now all the lines use the currency rate of the move.
This is a problem for upgrade. Instead of just (re)computing
the new column / field we use a different way here:
In a first query (Use the first method that is possible):
  1. Set it to 1 in some trivial cases
     (e.g. if the move currency is the company currency)
  2. Take it from fields that stored it already.
     (These fields are removed after this upgrade)
  3. Compute it from the ratio of the totals of the invoice
     in company and document currency
In a second query for all the remaining moves:
  4. Fetch the rate from the database (at invoice date).

There were already fields / modules with a similar purpose as the new field.
They were removed.

see odoo/odoo#147910
see odoo/enterprise#53609

task-3634224

closes odoo/upgrade#5496

Signed-off-by: Christophe Simonis (chs) <chs@odoo.com>