[FIX] hr_holidays/saas~16.5: optionally compute number of hours
The upgrade script to retroactively compute the number of hours for
existing leaves takes a significant amount of time. As this field is
currently only used for reporting and its correctness can only really
be guaranteed for leaves created after the upgrade, we added an option
to skip this upgrade altogether, or to only compute them for the
current calendar year (to keep ongoing leaves consistent). Non-upgrade
leaves will have their number of hours set to 0.

closes odoo/upgrade#5487

Signed-off-by: Christophe Simonis (chs) <chs@odoo.com>