[FIX] website/saas~16.5: expand update cowed view website.header_call_to_action
A fixup for https://github.com/odoo/upgrade/pull/5466/commits/4650b05db31b539adf7215cf2d14067f01fef438

Previous commit is modyfing a class attribute of a certain node. But these
cowed views were found to be sometimes missing the parent of that node altogether.

Therefore this commit aims to check for existence of that parent <section> element
and re-add it if need be:
https://github.com/odoo/odoo/blob/ad35458e8bfc30602f927944031a0bf87a92f82d/addons/website/views/website_templates.xml#L2185-L2189

It's crucial that the element fixed in the previous commit
exists, because in Odoo 17 there is a new view that is using it as an anchor:
https://github.com/odoo/odoo/blob/f1654683b68276955b55372aeee6139cbb75de1f/addons/website/views/website_templates.xml#L2196

opw-3644220
upg-1202958

closes odoo/upgrade#5500

Signed-off-by: Alvaro Fuentes Suarez (afu) <afu@odoo.com>