[FIX] base/saas~16.5: fix view attrs conv. when using field targetering
Considering the following custom view inherit on `project.task` model:

```
<field name="user_ids" position="after">
   <field name="other_field" attrs="{'invisible': [('project_id', 'not
in', [1, 2, 3])}"/>
</field>
```

The method `target_elem_and_view_type()` would return `user_ids` as part
of the field path of `other_field` field, which is incorrect as the node
will be added as a sibling, not as a children.

This result in wrong convertion of domain by the `convert_domain_leaf()`
as we fail to identify the proper field type.

In the preceding case, the 'invisible' domain was converted to:
  `(set(project_id).intersection([1, 2, 3]))`
instead of:
  `project_id not in [1, 2, 3]`

because the `project_id` field was wongly identifed as 'many2many' while
being a 'many2one' field.

This commit only consider targerting field with `position="inside"` as
being part of field path and fix field type identification when we
have a `field_path` (should be prepended)

closes odoo/upgrade#5512

Signed-off-by: Alvaro Fuentes Suarez (afu) <afu@odoo.com>