[FIX] l10n_tr_reports/saas~16.1: Remove account report expression record
There is an error related to a unique constraint in version saas-16.3.

In the version saas-16.2, introduced a new constraints.

Constraints named account_report_expression_line_label_uniq.

This constraint used in label column.

A problem occurs when this "account_report_tr_balance_sheet_data.xml" file.

Attempts to insert a value.

But a value with the same name already exists in that column.

This conflict results in an error.

closes odoo/upgrade#5511

Signed-off-by: Alvaro Fuentes Suarez (afu) <afu@odoo.com>