[FIX] sale_subscription/saas~16.5: plan created from archived quot.tmpl
Following odoo/upgrade@384195c33 we are creating new 'Plan' for each
qutoation template and recurrence - but if case the source quotation
template is archived, the plan created from it should be archived too.

This commit only set the new 'Plan' as active if both the recurrence
and the quotation template are active.