[FIX] l10n_mx_edi/saas~16.4.0.3: adapt mx edi cron for 16.4
This commit aims to add a cron adapted for changes introduced in 16.4.
In 16.4, all MX edi documents are migrated from `account.edi.document`
to `l10n_mx_edi.document`:
https://github.com/odoo/upgrade/blob/a1bf5aedaa4b267b4307f49255f17cc02f138bee/migrations/l10n_mx_edi/saas~16.4.0.3/post-migrate.py#L7

This means that in the existing cron, we have to adapt one of the SELECT
statements to account for any changes between the two models, such as
different selection values for the `state` field or that we may no
longer need to include edi_format_id in the WHERE clause (especially
that the edi_format "l10n_mx_edi.edi_cfdi_3_3" used prior
to 16.4 is removed when upgrading to 16.4).

closes odoo/upgrade#5532

Signed-off-by: Alvaro Fuentes Suarez (afu) <afu@odoo.com>